requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
asyncio-mqtt==0.16.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-sock==0.7.0
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0[p][\
]
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
openpyxl==3.1.5
paho-mqtt==2.1.0
pandas==2.3.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==0.19.0
pytz==2023.3.post1
requests==2.26.0
requests-oauthlib==2.0.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==1.4.46
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0